# api/openapi.yaml
openapi: 3.0.3
info:
  title: OCX Protocol API
  version: 1.0.0-rc.1
  description: Computational integrity through cryptographic receipts
  contact:
    name: OCX Protocol Support
    email: support@ocx.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.ocx.dev
    description: Production server

paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /execute:
    post:
      summary: Execute code and generate receipt
      operationId: executeCode
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
          description: Unique key to ensure idempotent execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteRequest'
      responses:
        '200':
          description: Execution successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimit'
        '500':
          $ref: '#/components/responses/InternalError'

  /verify:
    post:
      summary: Verify receipt cryptographically
      operationId: verifyReceipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
          application/cbor:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /receipts:
    get:
      summary: List receipts
      operationId: listReceipts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of receipts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptList'

  /keys/{id}:
    get:
      summary: Get key metadata
      operationId: getKeyMetadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Key metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyMetadata'

  /metrics:
    get:
      summary: Prometheus metrics
      operationId: getMetrics
      responses:
        '200':
          description: Prometheus metrics
          content:
            text/plain:
              schema:
                type: string

components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy]
        version:
          type: string
        uptime:
          type: string

    ExecuteRequest:
      type: object
      required: [artifact, input]
      properties:
        artifact:
          type: string
          format: base64
          description: Base64-encoded bytecode
        input:
          type: string
          format: base64
          description: Base64-encoded input data
        max_cycles:
          type: integer
          minimum: 1
          maximum: 1000000
          default: 10000

    ExecuteResponse:
      type: object
      properties:
        output_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        cycles_used:
          type: integer
        receipt_hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
        receipt_blob:
          type: string
          format: base64
        verify_command:
          type: string
          description: Ready-to-paste CLI verify command

    VerifyRequest:
      type: object
      required: [receipt]
      properties:
        receipt:
          type: string
          format: base64
          description: Base64-encoded CBOR receipt

    VerifyResponse:
      type: object
      properties:
        valid:
          type: boolean
        issuer_id:
          type: string
        cycles:
          type: integer
        timestamp:
          type: string
          format: date-time
        error:
          type: string

    ReceiptList:
      type: object
      properties:
        receipts:
          type: array
          items:
            $ref: '#/components/schemas/Receipt'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    Receipt:
      type: object
      properties:
        hash:
          type: string
        timestamp:
          type: string
          format: date-time
        cycles:
          type: integer
        issuer_id:
          type: string

    KeyMetadata:
      type: object
      properties:
        id:
          type: string
        public_key:
          type: string
        created_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, revoked]

    Error:
      type: object
      properties:
        code:
          type: string
          enum: [E001, E002, E003, E004, E005, E006]
        message:
          type: string
        details:
          type: object

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  code:
                    enum: [E001, E002]

    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  code:
                    enum: [E003]

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  code:
                    enum: [E004, E005, E006]